Add files via upload
Question 9. Create a general class ThreeDObject and derive the classes Box, Cube, Cylinder and 
Cone from it. The class ThreeDObject has methods wholeSurfaceArea ( ) and volume (). 
Override these two methods in each of the derived classes to calculate the volume and 
whole surface area of each type of three-dimensional objects. The dimensions of the 
objects are to be taken from the users and passed through the respective constructors of 
each derived class. Write a main method to test these classes.